--- apps/midscene-studio/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "midscene-studio",
-  "private": true,
-  "version": "0.1.0",
-  "type": "module",
-  "scripts": {
-    "build": "rsbuild build && node scripts/sync-static-assets.mjs",
-    "dev": "concurrently -k -n build,app \"rsbuild dev\" \"node scripts/wait-for-electron-build.mjs && MIDSCENE_STUDIO_RENDERER_URL=http://127.0.0.1:3210 electron dist/main/main.cjs\"",
-    "preview": "rsbuild preview --environment renderer",
-    "start": "node scripts/sync-static-assets.mjs && electron dist/main/main.cjs"
-  },
-  "dependencies": {
-    "react": "18.3.1",
-    "react-dom": "18.3.1"
-  },
-  "devDependencies": {
-    "@rsbuild/core": "^1.6.15",
-    "@rsbuild/plugin-less": "^1.5.0",
-    "@rsbuild/plugin-react": "^1.4.1",
-    "@rsbuild/plugin-type-check": "^1.3.2",
-    "@tailwindcss/postcss": "4.1.11",
-    "@types/node": "^18.0.0",
-    "@types/react": "^18.3.1",
-    "@types/react-dom": "^18.3.1",
-    "concurrently": "^8.2.0",
-    "electron": "41.2.0",
-    "less": "^4.2.0",
-    "tailwindcss": "4.1.11",
-    "typescript": "^5.8.3"
-  }
-}
