--- apps/midscene-studio/postcss.config.mjs
@@ -1,7 +1,0 @@
-export default {
-  plugins: {
-    '@tailwindcss/postcss': {
-      preflight: false,
-    },
-  },
-};
