--- apps/midscene-studio/src/renderer/App.css
@@ -1,29 +1,0 @@
-@layer theme, base, components, utilities;
-@import "tailwindcss/theme.css" layer(theme);
-@import "tailwindcss/utilities.css" layer(utilities);
-
-@layer base {
-  html,
-  body,
-  #root {
-    width: 100%;
-    height: 100%;
-    margin: 0;
-    overflow: hidden;
-    background: transparent;
-  }
-
-  body {
-    overscroll-behavior: none;
-  }
-}
-
-@layer utilities {
-  .app-drag {
-    -webkit-app-region: drag;
-  }
-
-  .app-no-drag {
-    -webkit-app-region: no-drag;
-  }
-}
