--- apps/midscene-studio/src/renderer/App.tsx
@@ -1,10 +1,0 @@
-import './App.css';
-import IncutAreaAllInOne from './components/incut-area-all-in-one';
-
-export default function App() {
-  return (
-    <div className="h-full w-full overflow-hidden bg-[#F6F6F6] text-slate-900">
-      <IncutAreaAllInOne />
-    </div>
-  );
-}
