+++ apps/studio/src/renderer/components/assets/main-disconnect.png
