+++ apps/studio/src/renderer/components/assets/main-phone-screen.png
