+++ apps/studio/src/renderer/components/assets/playground-action-chevron.png
