+++ apps/studio/src/renderer/components/assets/playground-history.png
