+++ apps/studio/src/renderer/components/assets/playground-send.png
