+++ apps/studio/src/renderer/components/assets/sidebar-harmony.png
