+++ apps/studio/src/renderer/components/assets/sidebar-ios.png
