+++ apps/studio/src/renderer/components/assets/sidebar-settings.png
