+++ apps/studio/src/renderer/components/assets/sidebar-web.png
