+++ apps/studio/src/renderer/components/globals.css
